+++ misc/coach.js
@@ -8,12 +8,14 @@ class Coach {
      * @param {int} lobbyCount 
      * @param {int} lobbyCountTryout 
      * @param {int} lobbyCountNormal
+     * @param {int} lobbyCountReplayAnalysis
      */
-    constructor(userId="", lobbyCount=0, lobbyCountTryout=0, lobbyCountNormal=0) {
+    constructor(userId="", lobbyCount=0, lobbyCountTryout=0, lobbyCountNormal=0, lobbyCountReplayAnalysis=0) {
         this.userId = userId
         this.lobbyCount = lobbyCount
         this.lobbyCountTryout = lobbyCountTryout
         this.lobbyCountNormal = lobbyCountNormal
+        this.lobbyCountReplayAnalysis = lobbyCountReplayAnalysis
     };
 
     static fromObject(obj) {