+++ misc/constants.js
@@ -1,6 +1,6 @@
 
-const lobbyTypes = {inhouse:1, unranked:2, botbash:3, tryout: 4}
-const lobbyTypePlayerCount = {inhouse:10, unranked:5, botbash:5, tryout: 5}
+const lobbyTypes = {inhouse:1, unranked:2, botbash:3, tryout: 4, replayAnalysis: 5}
+const lobbyTypePlayerCount = {inhouse:10, unranked:5, botbash:5, tryout: 5, replayAnalysis: 1000}
 const positionReactionEmojis = ['1️⃣', '2️⃣', '3️⃣', '4️⃣', '5️⃣'];
 const tryoutReactionEmoji = '✅';
 const lobbyManagementReactionEmojis =['🔒', '❌', '🧑‍🏫'];
@@ -10,8 +10,26 @@ module.exports = {
     positionReactionEmojis: positionReactionEmojis,
     lobbyManagementReactionEmojis: lobbyManagementReactionEmojis,
     lobbyTypes: lobbyTypes,
+    simpleLobbyTypes: [lobbyTypes.tryout, lobbyTypes.replayAnalysis],
+    roleBasedLobbyTypes: [lobbyTypes.inhouse, lobbyTypes.unranked, lobbyTypes.botbash],
     lobbyTypePlayerCount: lobbyTypePlayerCount,
     
+    /**
+     * Returns true if lobbytype is found in simpleLobbyTypes
+     * @param {number} lobbyType 
+     */
+    isSimpleLobbyType: function(lobbyType) {
+        return this.simpleLobbyTypes.find(s_type => s_type === lobbyType) !== undefined;
+    },
+
+    /**
+     * Returns true if lobbytype is found in roleBasedLobbyTypes
+     * @param {number} lobbyType 
+     */
+    isRoleBasedLobbyType: function(lobbyType) {
+        return this.roleBasedLobbyTypes.find(s_type => s_type === lobbyType) !== undefined;
+    },
+
     /**
      * Returns lobby name for usage in communication strings
      * @param {number} lobbyType given lobby type
@@ -28,6 +46,24 @@ module.exports = {
                 return "Botbash"
             case lobbyTypes.tryout:
                 return "Tryout"
+            case lobbyTypes.replayAnalysis:
+                return "Replay analysis"
+        }
+    },
+
+    getLobbyPostNameByType: function(lobbyType) {
+        switch (lobbyType)
+        {
+            case lobbyTypes.inhouse:
+                return "an INHOUSE lobby"
+            case lobbyTypes.unranked:
+                return "an UNRANKED lobby"
+            case lobbyTypes.botbash:
+                return "a BOTBASH lobby"
+            case lobbyTypes.tryout:
+                return "a TRYOUT lobby"
+            case lobbyTypes.replayAnalysis:
+                return "a REPLAY ANALYSIS session"
         }
     },
 
@@ -47,7 +83,7 @@ module.exports = {
         return positionReactionEmojis.includes(reactionEmoji.name);
     },
 
-    isKnownTryoutEmoji: function(reactionEmoji) {
+    isKnownSimpleLobbyEmoji: function(reactionEmoji) {
         return tryoutReactionEmoji === reactionEmoji.name;
     },